@@ -2,85 +2,183 @@ package cmd
 
 import (
 	"log"
+	"path/filepath"
 	"time"
 
+	"github.com/ErdemOzgen/blackdagger/internal/agent"
+	"github.com/ErdemOzgen/blackdagger/internal/client"
 	"github.com/ErdemOzgen/blackdagger/internal/config"
 	"github.com/ErdemOzgen/blackdagger/internal/dag"
-	"github.com/ErdemOzgen/blackdagger/internal/engine"
-	"github.com/ErdemOzgen/blackdagger/internal/persistence/client"
-	"github.com/ErdemOzgen/blackdagger/internal/scheduler"
+	"github.com/ErdemOzgen/blackdagger/internal/dag/scheduler"
+	"github.com/ErdemOzgen/blackdagger/internal/logger"
 	"github.com/spf13/cobra"
 )
 
 func restartCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "restart <DAG file>",
-		Short: "Restart the DAG",
-		Long:  `blackdagger restart <DAG file>`,
+	cmd := &cobra.Command{
+		Use:   "restart /path/to/spec.yaml",
+		Short: "Stop the running DAG and restart it",
+		Long:  `blackdagger restart /path/to/spec.yaml`,
 		Args:  cobra.ExactArgs(1),
-		PreRun: func(cmd *cobra.Command, args []string) {
-			cobra.CheckErr(config.LoadConfig())
-		},
 		Run: func(cmd *cobra.Command, args []string) {
-			dagFile := args[0]
-			loadedDAG, err := loadDAG(dagFile, "")
-			checkError(err)
+			cfg, err := config.Load()
+			if err != nil {
+				log.Fatalf("Configuration load failed: %v", err)
+			}
 
-			df := client.NewDataStoreFactory(config.Get())
-			e := engine.NewFactory(df, config.Get()).Create()
+			quiet, err := cmd.Flags().GetBool("quiet")
+			if err != nil {
+				log.Fatalf("Flag retrieval failed (quiet): %v", err)
+			}
 
-			// Check the current status and stop the DAG if it is running.
-			stopDAGIfRunning(e, loadedDAG)
+			initLogger := logger.NewLogger(logger.NewLoggerArgs{
+				Debug:  cfg.Debug,
+				Format: cfg.LogFormat,
+				Quiet:  quiet,
+			})
+
+			// Load the DAG file and stop the DAG if it is running.
+			specFilePath := args[0]
+			workflow, err := dag.Load(cfg.BaseConfig, specFilePath, "")
+			if err != nil {
+				initLogger.Fatal("Workflow load failed", "error", err, "file", args[0])
+			}
+
+			dataStore := newDataStores(cfg)
+			cli := newClient(cfg, dataStore, initLogger)
+
+			if err := stopDAGIfRunning(cli, workflow, initLogger); err != nil {
+				initLogger.Fatal("Workflow stop operation failed",
+					"error", err,
+					"workflow", workflow.Name)
+			}
 
 			// Wait for the specified amount of time before restarting.
-			waitForRestart(loadedDAG.RestartWait)
+			waitForRestart(workflow.RestartWait, initLogger)
 
 			// Retrieve the parameter of the previous execution.
-			log.Printf("Restarting %s...", loadedDAG.Name)
-			params := getPreviousExecutionParams(e, loadedDAG)
+			params, err := getPreviousExecutionParams(cli, workflow)
+			if err != nil {
+				initLogger.Fatal("Previous execution parameter retrieval failed",
+					"error", err,
+					"workflow", workflow.Name)
+			}
 
 			// Start the DAG with the same parameter.
-			loadedDAG, err = loadDAG(dagFile, params)
-			checkError(err)
-			cobra.CheckErr(start(cmd.Context(), e, loadedDAG, false))
+			// Need to reload the DAG file with the parameter.
+			workflow, err = dag.Load(cfg.BaseConfig, specFilePath, params)
+			if err != nil {
+				initLogger.Fatal("Workflow reload failed",
+					"error", err,
+					"file", specFilePath,
+					"params", params)
+			}
+
+			requestID, err := generateRequestID()
+			if err != nil {
+				initLogger.Fatal("Request ID generation failed", "error", err)
+			}
+
+			logFile, err := logger.OpenLogFile(logger.LogFileConfig{
+				Prefix:    "restart_",
+				LogDir:    cfg.LogDir,
+				DAGLogDir: workflow.LogDir,
+				DAGName:   workflow.Name,
+				RequestID: requestID,
+			})
+			if err != nil {
+				initLogger.Fatal("Log file creation failed",
+					"error", err,
+					"workflow", workflow.Name)
+			}
+			defer logFile.Close()
+
+			agentLogger := logger.NewLogger(logger.NewLoggerArgs{
+				Debug:   cfg.Debug,
+				Format:  cfg.LogFormat,
+				LogFile: logFile,
+				Quiet:   quiet,
+			})
+
+			agentLogger.Info("Workflow restart initiated",
+				"workflow", workflow.Name,
+				"requestID", requestID,
+				"logFile", logFile.Name())
+
+			agt := agent.New(
+				requestID,
+				workflow,
+				agentLogger,
+				filepath.Dir(logFile.Name()),
+				logFile.Name(),
+				newClient(cfg, dataStore, agentLogger),
+				dataStore,
+				&agent.Options{Dry: false})
+
+			listenSignals(cmd.Context(), agt)
+			if err := agt.Run(cmd.Context()); err != nil {
+				agentLogger.Fatal("Workflow restart failed",
+					"error", err,
+					"workflow", workflow.Name,
+					"requestID", requestID)
+			}
 		},
 	}
+	cmd.Flags().BoolP("quiet", "q", false, "suppress output")
+	return cmd
 }
 
-func stopDAGIfRunning(e engine.Engine, d *dag.DAG) {
-	st, err := e.GetCurrentStatus(d)
-	checkError(err)
-
-	// Stop the DAG if it is running.
-	if st.Status == scheduler.StatusRunning {
-		log.Printf("Stopping %s for restart...", d.Name)
-		cobra.CheckErr(stopRunningDAG(e, d))
+// stopDAGIfRunning stops the DAG if it is running.
+// Otherwise, it does nothing.
+func stopDAGIfRunning(e client.Client, workflow *dag.DAG, lg logger.Logger) error {
+	curStatus, err := e.GetCurrentStatus(workflow)
+	if err != nil {
+		return err
 	}
+
+	if curStatus.Status == scheduler.StatusRunning {
+		lg.Infof("Stopping: %s", workflow.Name)
+		cobra.CheckErr(stopRunningDAG(e, workflow))
+	}
+	return nil
 }
 
-func stopRunningDAG(e engine.Engine, d *dag.DAG) error {
+// stopRunningDAG attempts to stop the running DAG
+// by sending a stop signal to the agent.
+func stopRunningDAG(e client.Client, workflow *dag.DAG) error {
 	for {
-		st, err := e.GetCurrentStatus(d)
-		checkError(err)
+		curStatus, err := e.GetCurrentStatus(workflow)
+		if err != nil {
+			return err
+		}
 
-		if st.Status != scheduler.StatusRunning {
+		// If the DAG is not running, do nothing.
+		if curStatus.Status != scheduler.StatusRunning {
 			return nil
 		}
-		checkError(e.Stop(d))
+
+		if err := e.Stop(workflow); err != nil {
+			return err
+		}
+
 		time.Sleep(time.Millisecond * 100)
 	}
 }
 
-func waitForRestart(restartWait time.Duration) {
+// waitForRestart waits for the specified amount of time before restarting
+// the DAG.
+func waitForRestart(restartWait time.Duration, lg logger.Logger) {
 	if restartWait > 0 {
-		log.Printf("Waiting for %s...", restartWait)
+		lg.Info("Waiting for restart", "duration", restartWait)
 		time.Sleep(restartWait)
 	}
 }
 
-func getPreviousExecutionParams(e engine.Engine, d *dag.DAG) string {
-	st, err := e.GetLatestStatus(d)
-	checkError(err)
+func getPreviousExecutionParams(e client.Client, workflow *dag.DAG) (string, error) {
+	latestStatus, err := e.GetLatestStatus(workflow)
+	if err != nil {
+		return "", err
+	}
 
-	return st.Params
+	return latestStatus.Params, nil
 }
