@@ -1,41 +1,46 @@
 package cmd
 
 import (
-	"os"
 	"testing"
 	"time"
 
-	"github.com/ErdemOzgen/blackdagger/internal/scheduler"
+	"github.com/ErdemOzgen/blackdagger/internal/dag/scheduler"
+	"github.com/ErdemOzgen/blackdagger/internal/test"
 )
 
 func TestStopCommand(t *testing.T) {
-	tmpDir, _, ds := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
+	t.Run("StopDAG", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	dagFile := testDAGFile("stop.yaml")
+		dagFile := testDAGFile("long2.yaml")
 
-	// Start the DAG.
-	done := make(chan struct{})
-	go func() {
-		testRunCommand(t, startCmd(), cmdTest{args: []string{"start", dagFile}})
-		close(done)
-	}()
+		// Start the DAG.
+		done := make(chan struct{})
+		go func() {
+			testRunCommand(t, startCmd(), cmdTest{args: []string{"start", dagFile}})
+			close(done)
+		}()
 
-	time.Sleep(time.Millisecond * 100)
+		time.Sleep(time.Millisecond * 100)
 
-	// Wait for the DAG running.
-	// TODO: Do not use history store.
-	testLastStatusEventual(t, ds.NewHistoryStore(), dagFile, scheduler.StatusRunning)
+		// Wait for the DAG running.
+		testLastStatusEventual(
+			t,
+			setup.DataStore().HistoryStore(),
+			dagFile,
+			scheduler.StatusRunning,
+		)
 
-	// Stop the DAG.
-	testRunCommand(t, stopCmd(), cmdTest{
-		args:        []string{"stop", dagFile},
-		expectedOut: []string{"Stopping..."}})
+		// Stop the DAG.
+		testRunCommand(t, stopCmd(), cmdTest{
+			args:        []string{"stop", dagFile},
+			expectedOut: []string{"Stopping..."}})
 
-	// Check the last execution is cancelled.
-	// TODO: Do not use history store.
-	testLastStatusEventual(t, ds.NewHistoryStore(), dagFile, scheduler.StatusCancel)
-	<-done
+		// Check the last execution is cancelled.
+		testLastStatusEventual(
+			t, setup.DataStore().HistoryStore(), dagFile, scheduler.StatusCancel,
+		)
+		<-done
+	})
 }
