@@ -3,14 +3,22 @@ package dag
 import (
 	"errors"
 	"fmt"
-
-	"github.com/ErdemOzgen/blackdagger/internal/utils"
+	"os"
 )
 
-// Condition represents a condition to be evaluated by the agent.
+// Condition contains a condition and the expected value.
+// Conditions are evaluated and compared to the expected value.
+// The condition can be a command substitution or an environment variable.
+// The expected value must be a string without any substitutions.
 type Condition struct {
-	Condition string
-	Expected  string
+	Condition string // Condition to evaluate
+	Expected  string // Expected value
+}
+
+// eval evaluates the condition and returns the actual value.
+// It returns an error if the evaluation failed or the condition is invalid.
+func (c Condition) eval() (string, error) {
+	return substituteCommands(os.ExpandEnv(c.Condition))
 }
 
 var (
@@ -18,35 +26,37 @@ var (
 	errEvalCondition   = errors.New("failed to evaluate condition")
 )
 
-// Evaluate returns the actual value of the condition.
-func (c *Condition) Evaluate() (string, error) {
-	return utils.ParseVariable(c.Condition)
-}
-
-// CheckResult checks if the actual value of the condition matches the expected value.
-func (c *Condition) CheckResult(actualValue string) error {
-	if c.Expected != actualValue {
-		return fmt.Errorf("%w. Condition=%s Expected=%s Actual=%s", errConditionNotMet, c.Condition, c.Expected, actualValue)
+// evalCondition evaluates a single condition and checks the result.
+// It returns an error if the condition was not met.
+func evalCondition(c Condition) error {
+	actual, err := c.eval()
+	if err != nil {
+		return fmt.Errorf(
+			"%w. Condition=%s Error=%v", errEvalCondition, c.Condition, err,
+		)
 	}
+
+	if c.Expected != actual {
+		return fmt.Errorf(
+			"%w. Condition=%s Expected=%s Actual=%s",
+			errConditionNotMet,
+			c.Condition,
+			c.Expected,
+			actual,
+		)
+	}
+
 	return nil
 }
 
-// EvalCondition evaluates a single condition and checks the result.
-func EvalCondition(c *Condition) error {
-	actual, err := c.Evaluate()
-	if err != nil {
-		return fmt.Errorf("%w. Condition=%s Error=%v", errEvalCondition, c.Condition, err)
-	}
-	return c.CheckResult(actual)
-}
-
 // EvalConditions evaluates a list of conditions and checks the results.
-func EvalConditions(cond []*Condition) error {
+// It returns an error if any of the conditions were not met.
+func EvalConditions(cond []Condition) error {
 	for _, c := range cond {
-		err := EvalCondition(c)
-		if err != nil {
+		if err := evalCondition(c); err != nil {
 			return err
 		}
 	}
+
 	return nil
 }
