@@ -2,17 +2,22 @@ package jsondb
 
 import (
 	"bufio"
+	"encoding/json"
+	"errors"
 	"os"
-	"path"
-	"strings"
+	"path/filepath"
 	"sync"
 
 	"github.com/ErdemOzgen/blackdagger/internal/persistence/model"
-
-	"github.com/ErdemOzgen/blackdagger/internal/utils"
+	"github.com/ErdemOzgen/blackdagger/internal/util"
 )
 
-// Writer is the interface to write status to local file.
+var (
+	ErrWriterClosed  = errors.New("writer is closed")
+	ErrWriterNotOpen = errors.New("writer is not open")
+)
+
+// writer manages writing status to a local file.
 type writer struct {
 	target  string
 	dagFile string
@@ -22,38 +27,84 @@ type writer struct {
 	closed  bool
 }
 
-// Open opens the writer.
-func (w *writer) open() (err error) {
-	_ = os.MkdirAll(path.Dir(w.target), 0755)
-	w.file, err = utils.OpenOrCreateFile(w.target)
-	if err == nil {
-		w.writer = bufio.NewWriter(w.file)
+// open opens the writer.
+func (w *writer) open() error {
+	w.mu.Lock()
+	defer w.mu.Unlock()
+
+	if w.closed {
+		return ErrWriterClosed
 	}
-	return
+
+	if err := os.MkdirAll(filepath.Dir(w.target), 0755); err != nil {
+		return err
+	}
+
+	file, err := util.OpenOrCreateFile(w.target)
+	if err != nil {
+		return err
+	}
+
+	w.file = file
+	w.writer = bufio.NewWriter(file)
+	return nil
 }
 
-// Writer appends the status to the local file.
+// write appends the status to the local file.
 func (w *writer) write(st *model.Status) error {
 	w.mu.Lock()
 	defer w.mu.Unlock()
-	jsonb, _ := st.ToJson()
-	str := strings.ReplaceAll(string(jsonb), "\n", " ")
-	str = strings.ReplaceAll(str, "\r", " ")
-	_, err := w.writer.WriteString(str + "\n")
-	utils.LogErr("write status", err)
+
+	if w.closed {
+		return ErrWriterClosed
+	}
+
+	if w.writer == nil {
+		return ErrWriterNotOpen
+	}
+
+	jsonb, err := json.Marshal(st)
+	if err != nil {
+		return err
+	}
+
+	if _, err := w.writer.Write(jsonb); err != nil {
+		return err
+	}
+
+	if err := w.writer.WriteByte('\n'); err != nil {
+		return err
+	}
+
 	return w.writer.Flush()
 }
 
-// Close closes the writer.
-func (w *writer) close() (err error) {
+// close closes the writer.
+func (w *writer) close() error {
 	w.mu.Lock()
 	defer w.mu.Unlock()
-	if !w.closed {
-		err = w.writer.Flush()
-		utils.LogErr("flush file", err)
-		utils.LogErr("file sync", w.file.Sync())
-		utils.LogErr("file close", w.file.Close())
-		w.closed = true
+
+	if w.closed {
+		return nil
 	}
+
+	var err error
+	if w.writer != nil {
+		err = w.writer.Flush()
+	}
+
+	if w.file != nil {
+		if syncErr := w.file.Sync(); syncErr != nil && err == nil {
+			err = syncErr
+		}
+		if closeErr := w.file.Close(); closeErr != nil && err == nil {
+			err = closeErr
+		}
+	}
+
+	w.closed = true
+	w.writer = nil
+	w.file = nil
+
 	return err
 }
