@@ -3,18 +3,18 @@ package local
 import (
 	"errors"
 	"fmt"
+	"io/fs"
 	"os"
 	"path"
 	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/ErdemOzgen/blackdagger/internal/persistence/filecache"
-
 	"github.com/ErdemOzgen/blackdagger/internal/dag"
-	"github.com/ErdemOzgen/blackdagger/internal/grep"
 	"github.com/ErdemOzgen/blackdagger/internal/persistence"
-	"github.com/ErdemOzgen/blackdagger/internal/utils"
+	"github.com/ErdemOzgen/blackdagger/internal/persistence/filecache"
+	"github.com/ErdemOzgen/blackdagger/internal/persistence/grep"
+	"github.com/ErdemOzgen/blackdagger/internal/util"
 )
 
 type dagStoreImpl struct {
@@ -22,46 +22,35 @@ type dagStoreImpl struct {
 	metaCache *filecache.Cache[*dag.DAG]
 }
 
-func NewDAGStore(dir string) persistence.DAGStore {
-	ds := &dagStoreImpl{
-		dir:       dir,
-		metaCache: filecache.New[*dag.DAG](0, time.Hour*24),
-	}
-	ds.metaCache.StartEviction()
-	return ds
+type NewDAGStoreArgs struct {
+	Dir string
 }
 
-var (
-	errInvalidName           = errors.New("invalid name")
-	errFailedToReadDAGFile   = errors.New("failed to read DAG file")
-	errDOGFileNotExist       = errors.New("the DAG file does not exist")
-	errFailedToUpdateDAGFile = errors.New("failed to update DAG file")
-	errFailedToCreateDAGFile = errors.New("failed to create DAG file")
-	errFailedToCreateDAGsDir = errors.New("failed to create DAGs directory")
-	errFailedToDeleteDAGFile = errors.New("failed to delete DAG file")
-	errDAGFileAlreadyExists  = errors.New("the DAG file already exists")
-	errInvalidNewName        = errors.New("invalid new name")
-	errInvalidOldName        = errors.New("invalid old name")
-)
+func NewDAGStore(args *NewDAGStoreArgs) persistence.DAGStore {
+	dagStore := &dagStoreImpl{
+		dir:       args.Dir,
+		metaCache: filecache.New[*dag.DAG](0, time.Hour*24),
+	}
+	dagStore.metaCache.StartEviction()
+	return dagStore
+}
 
 func (d *dagStoreImpl) GetMetadata(name string) (*dag.DAG, error) {
 	loc, err := d.fileLocation(name)
 	if err != nil {
-		return nil, fmt.Errorf("%w: %s", errInvalidName, name)
+		return nil, err
 	}
 	return d.metaCache.LoadLatest(loc, func() (*dag.DAG, error) {
-		cl := dag.Loader{}
-		return cl.LoadMetadata(loc)
+		return dag.LoadMetadata(loc)
 	})
 }
 
 func (d *dagStoreImpl) GetDetails(name string) (*dag.DAG, error) {
 	loc, err := d.fileLocation(name)
 	if err != nil {
-		return nil, fmt.Errorf("%w: %s", errInvalidName, name)
+		return nil, err
 	}
-	cl := dag.Loader{}
-	dat, err := cl.LoadWithoutEval(loc)
+	dat, err := dag.LoadWithoutEval(loc)
 	if err != nil {
 		return nil, err
 	}
@@ -71,59 +60,66 @@ func (d *dagStoreImpl) GetDetails(name string) (*dag.DAG, error) {
 func (d *dagStoreImpl) GetSpec(name string) (string, error) {
 	loc, err := d.fileLocation(name)
 	if err != nil {
-		return "", fmt.Errorf("%w: %s", errInvalidName, name)
+		return "", err
 	}
 	dat, err := os.ReadFile(loc)
 	if err != nil {
-		return "", fmt.Errorf("%w: %s", errFailedToReadDAGFile, err)
+		return "", err
 	}
 	return string(dat), nil
 }
 
+// TODO: use 0600 // nolint: gosec
+const defaultPerm os.FileMode = 0744
+
+var errDOGFileNotExist = errors.New("the DAG file does not exist")
+
 func (d *dagStoreImpl) UpdateSpec(name string, spec []byte) error {
 	// validation
-	cl := dag.Loader{}
-	_, err := cl.LoadData(spec)
+	_, err := dag.LoadYAML(spec)
 	if err != nil {
 		return err
 	}
 	loc, err := d.fileLocation(name)
 	if err != nil {
-		return fmt.Errorf("%w: %s", errInvalidName, name)
+		return err
 	}
 	if !exists(loc) {
 		return fmt.Errorf("%w: %s", errDOGFileNotExist, loc)
 	}
-	err = os.WriteFile(loc, spec, 0755)
+	err = os.WriteFile(loc, spec, defaultPerm)
 	if err != nil {
-		return fmt.Errorf("%w: %s", errFailedToUpdateDAGFile, err)
+		return err
 	}
 	d.metaCache.Invalidate(loc)
 	return nil
 }
 
+var errDAGFileAlreadyExists = errors.New("the DAG file already exists")
+
 func (d *dagStoreImpl) Create(name string, spec []byte) (string, error) {
 	if err := d.ensureDirExist(); err != nil {
-		return "", fmt.Errorf("%w: %s", errFailedToCreateDAGsDir, d.dir)
+		return "", err
 	}
 	loc, err := d.fileLocation(name)
 	if err != nil {
-		return "", fmt.Errorf("%w: %s", errFailedToCreateDAGFile, name)
+		return "", err
 	}
 	if exists(loc) {
 		return "", fmt.Errorf("%w: %s", errDAGFileAlreadyExists, loc)
 	}
+	// nolint: gosec
 	return name, os.WriteFile(loc, spec, 0644)
 }
 
 func (d *dagStoreImpl) Delete(name string) error {
 	loc, err := d.fileLocation(name)
 	if err != nil {
-		return fmt.Errorf("%w: %s", errFailedToCreateDAGFile, name)
+		return err
 	}
 	err = os.Remove(loc)
 	if err != nil {
-		return fmt.Errorf("%w: %s", errFailedToDeleteDAGFile, err)
+		return err
 	}
 	d.metaCache.Invalidate(loc)
 	return nil
@@ -139,14 +135,7 @@ func (d *dagStoreImpl) fileLocation(name string) (string, error) {
 		// this is for backward compatibility
 		return name, nil
 	}
-	loc := path.Join(d.dir, name)
-	return d.normalizeFilename(loc)
-}
-
-func (d *dagStoreImpl) normalizeFilename(file string) (string, error) {
-	a := strings.TrimSuffix(file, ".yaml")
-	a = strings.TrimSuffix(a, ".yml")
-	return fmt.Sprintf("%s.yaml", a), nil
+	return util.AddYamlExtension(path.Join(d.dir, name)), nil
 }
 
 func (d *dagStoreImpl) ensureDirExist() error {
@@ -156,6 +145,84 @@ func (d *dagStoreImpl) ensureDirExist() error {
 		}
 	}
 	return nil
+}
+
+func (d *dagStoreImpl) searchName(fileName string, searchText *string) bool {
+	if searchText == nil {
+		return true
+	}
+
+	fileName = strings.TrimSuffix(fileName, path.Ext(fileName))
+
+	return strings.Contains(fileName, *searchText)
+}
+
+func (d *dagStoreImpl) searchTags(tags []string, searchTag *string) bool {
+	if searchTag == nil {
+		return true
+	}
+
+	for _, tag := range tags {
+		if tag == *searchTag {
+			return true
+		}
+	}
+
+	return false
+}
+
+func (d *dagStoreImpl) getTagList(tagSet map[string]struct{}) []string {
+	tagList := make([]string, 0, len(tagSet))
+	for tag := range tagSet {
+		tagList = append(tagList, tag)
+	}
+	return tagList
+}
+
+func (d *dagStoreImpl) ListPagination(params persistence.DAGListPaginationArgs) (*persistence.DagListPaginationResult, error) {
+	var (
+		dagList    = make([]*dag.DAG, 0)
+		errList    = make([]string, 0)
+		count      int
+		currentDag *dag.DAG
+	)
+
+	if err := filepath.WalkDir(d.dir, func(_ string, dir fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+
+		if dir.IsDir() || !checkExtension(dir.Name()) {
+			return nil
+		}
+
+		if currentDag, err = d.GetMetadata(dir.Name()); err != nil {
+			errList = append(errList, fmt.Sprintf("reading %s failed: %s", dir.Name(), err))
+		}
+
+		if !d.searchName(dir.Name(), params.Name) || currentDag == nil || !d.searchTags(currentDag.Tags, params.Tag) {
+			return nil
+		}
+
+		count++
+		if count > (params.Page-1)*params.Limit && len(dagList) < params.Limit {
+			dagList = append(dagList, currentDag)
+		}
+
+		return nil
+	}); err != nil {
+		return &persistence.DagListPaginationResult{
+			DagList:   dagList,
+			Count:     count,
+			ErrorList: append(errList, err.Error()),
+		}, err
+	}
+
+	return &persistence.DagListPaginationResult{
+		DagList:   dagList,
+		Count:     count,
+		ErrorList: errList,
+	}, nil
 }
 
 func (d *dagStoreImpl) List() (ret []*dag.DAG, errs []string, err error) {
@@ -174,7 +241,9 @@ func (d *dagStoreImpl) List() (ret []*dag.DAG, errs []string, err error) {
 			if err == nil {
 				ret = append(ret, dat)
 			} else {
-				errs = append(errs, fmt.Sprintf("reading %s failed: %s", fi.Name(), err))
+				errs = append(errs, fmt.Sprintf(
+					"reading %s failed: %s", fi.Name(), err),
+				)
 			}
 		}
 	}
@@ -193,42 +262,49 @@ func checkExtension(file string) bool {
 	return false
 }
 
-func (d *dagStoreImpl) Grep(pattern string) (ret []*persistence.GrepResult, errs []string, err error) {
+func (d *dagStoreImpl) Grep(
+	pattern string,
+) (ret []*persistence.GrepResult, errs []string, err error) {
 	if err = d.ensureDirExist(); err != nil {
-		errs = append(errs, fmt.Sprintf("failed to create DAGs directory %s", d.dir))
+		errs = append(
+			errs, fmt.Sprintf("failed to create DAGs directory %s", d.dir),
+		)
 		return
 	}
 
 	fis, err := os.ReadDir(d.dir)
-	dl := &dag.Loader{}
 	opts := &grep.Options{
 		IsRegexp: true,
 		Before:   2,
 		After:    2,
 	}
 
-	utils.LogErr("read DAGs directory", err)
+	util.LogErr("read DAGs directory", err)
 	for _, fi := range fis {
-		if utils.MatchExtension(fi.Name(), dag.EXTENSIONS) {
+		if util.MatchExtension(fi.Name(), dag.Exts) {
 			file := filepath.Join(d.dir, fi.Name())
 			dat, err := os.ReadFile(file)
 			if err != nil {
-				utils.LogErr("read DAG file", err)
+				util.LogErr("read DAG file", err)
 				continue
 			}
 			m, err := grep.Grep(dat, fmt.Sprintf("(?i)%s", pattern), opts)
 			if err != nil {
-				errs = append(errs, fmt.Sprintf("grep %s failed: %s", fi.Name(), err))
+				errs = append(
+					errs, fmt.Sprintf("grep %s failed: %s", fi.Name(), err),
+				)
 				continue
 			}
-			d, err := dl.LoadMetadata(file)
+			dg, err := dag.LoadMetadata(file)
 			if err != nil {
-				errs = append(errs, fmt.Sprintf("check %s failed: %s", fi.Name(), err))
+				errs = append(
+					errs, fmt.Sprintf("check %s failed: %s", fi.Name(), err),
+				)
 				continue
 			}
 			ret = append(ret, &persistence.GrepResult{
 				Name:    strings.TrimSuffix(fi.Name(), path.Ext(fi.Name())),
-				DAG:     d,
+				DAG:     dg,
 				Matches: m,
 			})
 		}
@@ -236,36 +312,30 @@ func (d *dagStoreImpl) Grep(pattern string) (ret []*persistence.GrepResult, errs
 	return ret, errs, nil
 }
 
-func (d *dagStoreImpl) Load(name string) (*dag.DAG, error) {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (d *dagStoreImpl) Rename(oldDAGPath, newDAGPath string) error {
-	oldLoc, err := d.fileLocation(oldDAGPath)
+func (d *dagStoreImpl) Rename(oldID, newID string) error {
+	oldLoc, err := d.fileLocation(oldID)
 	if err != nil {
-		return fmt.Errorf("%w: %s", errInvalidOldName, oldDAGPath)
+		return err
 	}
-	newLoc, err := d.fileLocation(newDAGPath)
+	newLoc, err := d.fileLocation(newID)
 	if err != nil {
-		return fmt.Errorf("%w: %s", errInvalidNewName, newDAGPath)
+		return err
 	}
 	return os.Rename(oldLoc, newLoc)
 }
 
-func (d *dagStoreImpl) FindByName(name string) (*dag.DAG, error) {
+func (d *dagStoreImpl) Find(name string) (*dag.DAG, error) {
 	file, err := d.resolve(name)
 	if err != nil {
 		return nil, err
 	}
-	cl := dag.Loader{}
-	return cl.Load(file, "")
+	return dag.LoadWithoutEval(file)
 }
 
 func (d *dagStoreImpl) resolve(name string) (string, error) {
 	// check if the name is a file path
 	if strings.Contains(name, string(filepath.Separator)) {
-		if !utils.FileExists(name) {
+		if !util.FileExists(name) {
 			return "", fmt.Errorf("workflow %s not found", name)
 		}
 		return name, nil
@@ -298,14 +368,51 @@ func find(name string) (string, error) {
 	ext := path.Ext(name)
 	if ext == "" {
 		// try all supported extensions
-		for _, ext := range dag.EXTENSIONS {
-			if utils.FileExists(name + ext) {
+		for _, ext := range dag.Exts {
+			if util.FileExists(name + ext) {
 				return filepath.Abs(name + ext)
 			}
 		}
-	} else if utils.FileExists(name) {
+	} else if util.FileExists(name) {
 		// the name has an extension
 		return filepath.Abs(name)
 	}
 	return "", fmt.Errorf("sub workflow %s not found", name)
+}
+
+func (d *dagStoreImpl) TagList() ([]string, []string, error) {
+	var (
+		errList    = make([]string, 0)
+		tagSet     = make(map[string]struct{})
+		currentDag *dag.DAG
+		err        error
+	)
+
+	if err = filepath.WalkDir(d.dir, func(_ string, dir fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+
+		if dir.IsDir() || !checkExtension(dir.Name()) {
+			return nil
+		}
+
+		if currentDag, err = d.GetMetadata(dir.Name()); err != nil {
+			errList = append(errList, fmt.Sprintf("reading %s failed: %s", dir.Name(), err))
+		}
+
+		if currentDag == nil {
+			return nil
+		}
+
+		for _, tag := range currentDag.Tags {
+			tagSet[tag] = struct{}{}
+		}
+
+		return nil
+	}); err != nil {
+		return nil, append(errList, err.Error()), err
+	}
+
+	return d.getTagList(tagSet), errList, nil
 }
